[#722] 반복옵션 요약 텍스트를 CommonPresentation으로 승격 — SelectRepeatingOptionModel이 공용 API에 위임
위젯 익스텐션은 EventDetailScene을 의존하지 않아 반복 주기 요약("매주 월" 등)을 만들 수 없었다. 변환 로직을 위젯이 import 가능한 모듈로 올린다.

- EventRepeatingOption.summaryText(startTime:timeZone:) — 신규. 반복옵션을 한 줄 요약으로, 미지원 조합은 nil
- SupportingOptions — EventDetailScene private 정의에서 CommonPresentation으로 이동 (텍스트 매핑과 한 단위)
- SelectRepeatingOptionModel.init?(_:_:_:) — 70줄 switch를 summaryText 위임 3줄로 축소, 동작 동일
- SupportedRepeatOptions — 텍스트 변환과 무관한 선택 후보 팩토리(supports)만 남긴 네임스페이스로 분리
